Refactor testing and database into separate modules.
