+++ modules/chat/server-ts/sql.ts
@@ -1,6 +1,5 @@
 /*tslint:disable:variable-name*/
-import { orderedFor, returnId } from '../../../packages/server/src/sql/helpers';
-import knex from '../../../packages/server/src/sql/connector';
+import { knex, orderedFor, returnId } from '@module/database-server-ts';
 import { UploadedFile } from '../../../packages/server/src/modules/upload/FileSystemStorage';
 
 export interface Identifier {
