+++ modules/database/server-ts/seeds/001_counter.js
@@ -1,4 +1,4 @@
-import { returnId, truncateTables } from '../../sql/helpers';
+import { returnId, truncateTables } from '../sql/helpers';
 
 const initialAmount = 5;
