+++ modules/database/server-ts/seeds/003_post.js
@@ -1,4 +1,4 @@
-import { returnId, truncateTables } from '../../sql/helpers';
+import { returnId, truncateTables } from '../sql/helpers';
 
 export async function seed(knex, Promise) {
   await truncateTables(knex, Promise, ['post', 'comment']);
