+++ modules/database/server-ts/sql/crud.js
@@ -1,14 +1,13 @@
 import _ from 'lodash';
 import uuidv4 from 'uuid';
 import { camelize, decamelizeKeys, camelizeKeys } from 'humps';
+import { log } from '@module/core-common';
 
 import knex from './connector';
 
 import { orderedFor } from './helpers';
 
 import selectAdapter from './select';
-
-import log from '../../../common/log';
 
 export function createWithIdGenAdapter(options) {
   const T = options.table;
