+++ modules/database/server-ts/sql/index.ts
@@ -1,0 +1,3 @@
+export { default as knex } from './connector';
+export { default as populateTestDb } from './populateTestDb';
+export * from './helpers';
