+++ modules/database/server-ts/sql/populateTestDb.js
@@ -1,0 +1,6 @@
+import knex from './connector';
+
+export default async () => {
+  await knex.migrate.latest();
+  return knex.seed.run();
+};
