+++ modules/testing/server-ts/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@module/testing-server-ts",
+  "version": "1.0.0",
+  "private": true
+}
