+++ packages/server/package.json
@@ -57,6 +57,8 @@
     "@babel/polyfill": "^7.0.0",
     "@babel/preset-env": "^7.0.0",
     "@babel/register": "^7.0.0",
+    "@module/database-server-ts": "^1.0.0",
+    "@module/testing-server-ts": "^1.0.0",
     "apollo-cache-inmemory": "1.2.9",
     "apollo-cache-router": "^1.1.6",
     "apollo-client": "^2.4.7",
