+++ packages/server/src/modules/user/testHelpers/index.js
@@ -1,7 +1,7 @@
 import LOGIN from '@module/user-client-react/graphql/Login.graphql';
 import LOGOUT from '@module/user-client-react/access/session/graphql/Logout.graphql';
 
-import { getApollo } from '../../../testHelpers/integrationSetup';
+import { getApollo } from '@module/testing-server-ts';
 
 export const login = async (usernameOrEmail = 'admin', password = 'admin123') =>
   await getApollo().mutate({
