+++ packages/server/src/sql/connector.ts
@@ -1,0 +1,2 @@
+// Reexport for backward-compatibility purposes
+export { knex as default } from '@module/database-server-ts';
